HOME-189 - Fix automated agents refactoring bugs
