+++ client/modules/Dashboard/TemperaturePanel/TemperatureChart/constants.js
@@ -1,0 +1,4 @@
+export const CHART_PADDING = 20;
+export const DEFAULT_MAX_TEMP = 30;
+export const DEFAULT_MIN_TEMP = 10;
+export const TEMP_MARGIN = 2;
